ci: separate oci-pull tests from legacy tests
since they're considered integration tests and can't be run inside
a docker container due to testcontainers-go
